Create comparison table at end of implementation tests.
Modified all implementation functions to return a float elapsed time
for the current test size. This is saved in a float array created in
main, which holds all elapsed times returned from the three variants.
This is presented later in a table, so that results are easily compared.

+++ cublas.h
@@ -1,6 +1,6 @@
 #ifndef CUBLAS_H
 #define CUBLAS_H
 
-int runCublas(int rows, int cols);
+float runCublas(int rows, int cols);
 
 #endif // CUBLAS_H
