+++ plainKernel.h
@@ -1,6 +1,6 @@
 #ifndef PLAINKERNEL_H
 #define PLAINKERNEL_H
 
-int plainKernelSetup(int, int, bool = false);
+float plainKernelSetup(int, int, bool = false);
 
 #endif
